ExprNarrowing: Correct the logic and clean up the implementation.
